requirements.txt
streamlit
speechrecognition
pyttsx3
python-dotenv
langchain-community
langchain-google-genai
langchain-pinecone
langchain-core
pinecone-client